Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.21.1
pymongo==4.8.0
python-dotenv
passlib==1.7.4
jwt==1.3.1
# bson
sendgrid==6.11.0
jinja2== 3.1.4
starlette==0.37.2
authlib==1.4.0
itsdangerous== 2.2.0
pydantic==2.7.4
httpx==0.24.1
cloudinary==1.44.0